General cleanup, setup auth schema
This is going to use a grants system, which is granted to a
particular password authentication. Every user can have multiple auths,
but only one "human auth". To login with the "human" auth, they must
prove they are not a robot (i.e., fill a recaptcha).

Note that these only correspond to the "permanent" grants, ie., grants
which don't depend on the state of a third-party resource. For grants
which are earned from meeting certain requirements, such as being an
approved submitter, the results are at best cached in an ephemeral
location.

There aren't many logins to redditloans so the current plan is 1m reps
of sha512 using pbkdf2_hmac for human logins, and 100k of sha512
for bot logins.