Add loan tables (#14)
* Add loan tables

A future commit (when adding the schemas in docs) will go into more
depth, but the general idea here is that the "loans" table contains
everything we need for knowing the *current state* of a single loan.
For historical information on a loan, one has to look at the
loan_*_events tables.

There are only 3 things that can happen to a loan right now after it is
created - some amount of principal can be marked repaid, the loan can
be marked delinquent, and an admin can change it. If you know the
current state of the loan and all the events on it, you can walk
backward to the state of the loan at any previous time.

Caution was taken with foreign keys here - the nature of moneys/
currencies tables is that it almost never makes sense to look up a
loan by the money amount - so despite a foreign key there, there
is no need for an index. This means that _deleting_ moneys or
currencies rows is expensive (linear with overall number of loans),
but there shouldn't be any reason to _ever_ do that.

* Allow 5 characters for symbol

Just in case a currencies symbol is not well handled in UTF-8 or
by most browsers, we can use the code + a space. All official
iso 4217 codes are 3 characters, but there some unofficial
4-character codes already so there's a reasonable chance they
one eventually become standard.

* Update docs for new loans tables